requirements.txt
aioboto3==15.5.0
aiobotocore==2.25.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aioitertools==0.13.0
aiosignal==1.4.0
alembic==1.14.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
billiard==4.2.2
boto3==1.40.61
botocore==1.40.61
cachetools==5.5.2
celery==5.4.0
certifi==2025.11.12
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
distlib==0.4.0
fastapi==0.115.4
filelock==3.20.0
frozenlist==1.8.0
google-api-core==2.28.1
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
h11==0.16.0
html2text==2020.1.16
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.27.2
identify==2.6.15
idna==3.11
iniconfig==2.3.0
jmespath==1.0.1
kombu==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
multidict==6.7.0
nodeenv==1.9.1
packaging==25.0
platformdirs==4.5.0
pluggy==1.6.0
pre_commit==4.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.2.5
pytest==8.3.3
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==5.0.8
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uv==0.9.9
uvicorn==0.32.0
vine==5.1.0
virtualenv==20.35.4
watchfiles==1.1.1
wcwidth==0.2.14
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
